Simplifying the pairs matching method.
