+++ CuriousKatie/Helper/Helper.swift
@@ -105,14 +105,4 @@ struct Helper {
         let duplicates = Dictionary(grouping: people, by: { $0.name }).filter { $1.count > 1 }.keys
         return (duplicates.count >= 1)
     }
-    
-    /// Calculating the matching score between two people by counting the number of different interests that they have
-    /// - Parameter seeker: the Person seeking for a partner
-    /// - Parameter potentialPartner: the potential partner Person
-    static func calculateMatchingScore(between seeker: Person, and potentialPartner: Person) -> Int {
-        let seekersInterests = seeker.interests.map { $0.name }
-        let potentialPartnersInterests = potentialPartner.interests.map { $0.name }
-        let differentInterests = seekersInterests.difference(from: potentialPartnersInterests)
-        return differentInterests.count
-    }
 }
